Revert "Move article components into page dir"
This reverts commit 9b215544b2277c171f488df8867e727fe12631da.
