+++ src/components2/Article/BarChart.tsx
