+++ src/components2/Article/BarChartWrapper.tsx
