+++ src/components2/Article/Quote.tsx
