+++ src/components2/Article/Text.tsx
